Task estimatedTime is now double ree

+++ src/main/java/dk/aau/ds304e18/database/DatabaseTask.java
@@ -19,7 +19,7 @@ class DatabaseTask {
     /**
      * The estimated completion of the task.
      */
-    public int estimatedTime;
+    public double estimatedTime;
 
     /**
      * The list of ids of employees assigned to the task.
